.github: Add eclass/usr-ldscript.eclass to automation
It couldn't be added before to automation, because the eclass in
Gentoo would introduce have unwanted side-effects into our built
images. But the Gentoo version of the eclass is essentially a no-op
when USE=split-usr is disabled. We have recently moved to use Gentoo
profiles that disable USE=split-usr altogether, and with this move, we
can now safely put the eclass into automation.

This marks the point where the entirety of the portage-stable is under
automation.