Drop --root-deps option from board emerge wrappers
This option used to install build dependencies to ROOT instead of /.
This never made much sense, so the option was rendered ineffective from
EAPI 7. The number of ebuilds with older EAPIs has since dwindled to
nothing.

A recent Portage change has made this option now install build
dependencies to ROOT as well as / because this can actually be useful
and doesn't cause breakage.

However, it does make us more prone to cyclic dependencies when
initially populating the board roots. There is no reason for us to use
this option though because its main purpose is to ensure the target
environment has everything it needs to rebuild itself. Given that the
option didn't do anything recently, we evidently don't even require this
for the developer container.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>