build_sysext, sysext_prod_builder: Fix mounting overlayfs with new util-linux
It is not really a fix, it is more of a workaround. I think that the
issue is either in util-linux or in the Linux kernel.

For some reason we can't mount the overlayfs with the new API. This
only happens when building the docker sysext, which means that
containerd pkginfo squashfs is also a part of the overlay lowerdirs. I
have no idea if this fact is even relevant.