sys-kernel/coreos-modules: Enable KERNEL_ZBOOT for arm64
Arm64 now supports booting a compressed kernel through EFI. This is
hidden behind the KERNEL_ZBOOT Kconfig option. The KERNEL_ZBOOT is only
enabled in the arm64 config because it is not available for x86. X86
relies on an architecture specific compression mechanism on EFI. The
KERNEL_ZSTD Kconfig option applies to both arches so move it to
commonconfig. The other change required is that the kernel image name
changes.

In my local build it looks like the kernel size goes down from 59MB to
43MB.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>