sys-apps/systemd: remove the networkd-default-to-kernel-IPForwarding-setting.patch
According to https://github.com/systemd/systemd-stable/blob/v256/src/network/networkd-network.c#L470,
the forwarding settings have changed on systemd 256.

From the discussions upstream, if a systemd is configured to manage an interface,
it will manage it completely, and it will set that interface to not forward packets
by default.

From the current systemd code, it would be easy to either enable the forwarding or disable it,
but there does not seem to be a way now to inherit it from the sysctl / kernel implementation.