run_sdk_container: Bail out on unknown flags
The script would pass unknown flags further as a container
command. This normally is not desired, but in case it actually is, the
user can use the newly added `--` parameter to stop parameter handling
and pass the following parameters as the container command.

I was tripped by using -u instead of -U and I got a rather confusing
error message suggesting something going wrong inside the container.