pkg-auto: Emerge dev-lang/rust for SDK reports
The coreos-devel/sdk-depends is a metapackage that is emerged in
stage4 of the catalyst SDK build. But the fsscript that is being run
at the end of the stage4 build also pulls in dev-lang/rust package.
Thus pull both packages for the report.

This should result in updated dev-lang/rust to appear in the reports.