ci-automation/image-changes: Fix fetching LTS version
When cutting a new LTS major release, we should fetch the current LTS
major release. For that, using the "current" symlink is enough -
fetching "lts-info" won't give us good results (either there won't be
an entry for the new major version or there won't be a
"current-${year}" symlink yet to follow).