install-nvidia: Remove depmod call
This was ineffective because that part of the container image is not mounted
mutable, so the depmod files were missing the nvidia kmods. Generating the
depmod files in a devcontainer also conflicts with depmod information for kmod
sysexts like zfs.

Since we load the module manually with insmod, we did not depend on the depmod
and it can be removed.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>