setup-nvidia: Modprobe video.ko for amd64
Since we added CONFIG_I915 to our amd64 kernel, CONFIG_ACPI_VIDEO=m is enabled
and causes the nvidia driver to compile in some code that depends on it. Since
we don't have depmod information and we insmod the nvidia module - manually
load the required module. The '|| true' lets this line pass on arm64.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>