setup-nvidia: Overlay host /lib/modules into devcontainer
So that we can pick-up kmods contained in sysexts (like zfs) and generate
complete module dependency information. I thought we could skip running depmod
for nvidia drivers because we manually insmod them, but nvidia's GPU operator
driver validation expects to be able to run modprobe - so we have to generate
them.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>