app-crypt/clevis: Version bump to 21
This moves the systemd unit enabling to the image build scripts to make
the ebuild less Flatcar-specific.

Unfortunately, Clevis is still very automagic, resulting in a poor
quality ebuild. Improving this was actually the very first thing I tried
to do for Flatcar back in 2022, 1½ years before I joined the team. I
will try to revive this effort soon, and then we can maybe get the
package upstreamed to Gentoo.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>