sys-kernel/coreos-kernel: Adapt for new Dracut + bootengine with sysroot
Dracut includes native binaries that are executed during the creation of
the initrd, so it always needs to be installed to /. For now, it expects
to find its modules in the sysroot, so it needs to be installed there
too, though that may change.

bootengine only needs to be installed to sysroot for its Dracut modules.
The update-bootengine tool is a Bash script, so it doesn't really need
to be installed to /.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>