ci-automation: Set the right parameters for Hyper-V Gen 1/2 on Azure
We still perform some tests using Gen 1 on amd64. Standard_D2s_v6 does
not support this, but v5 will presumably be the last version that does,
so hardcode that case.

For Gen 2, you need to set the SKU for amd64 to work, and it has to use
the gallery like arm64 already does.

Using the gallery is possibly slightly slower, so ideally we would only
upload the image once like we do for AWS, but let's just get it working
for now.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>