sys-firmware/intel-microcode: Fix inclusion of microcode in the initrd
The microcode was accidentally dropped from Flatcar a while back because
Gentoo started telling Dracut to not include it when USE=-initramfs.
Flatcar disabled that flag because the microcode is installed to /boot.
This is only done under the board root though, not in the final image,
so there's no harm in enabling it.

That Dracut setting also affected the AMD microcode, which is part of
coreos-firmware, so this fixes that too.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>