overlay app-emulation/wa-linux-agent: Stop using legacy distutils workflow
The legacy workflow was deprecated for some time in distutils-r1
eclass and was recently removed. Now specifying the
DISTUTILS_USE_PEP517 variable is mandatory, but there is no distutils
workflow that would match what legacy was doing. The closest ones are
either "no" or "setuptools". I picked "setuptools", but it results in
all the files put inside the site-packages directory. As a workaround,
move some of the files back where they were supposed to be.

While at it, regenerate the patches and update the python
compatibility variable.