build_library/qemu_template.sh: exec qemu instead of calling it
This change updates the build library's qemu template to "exec" qemu instead of calling it from the script. While serving the same purpose as calling qemu and returning "$?" (the current implementation), using exec makes it easier to manage the qemu sub-process when the qemu wrapper script is called from another script.