pkg-auto: Fix generation of reports of key-value pairs
It was a grouping issue where we treated spaces before key-value pairs
as not optional. Fixing this by making them optional created another
issue, where we ran out of capture groups - sed support only 10
capture groups from \0 to \9, and the first fix caused the keyvals to
become 11th captured group. To work around this issue, I added some
non-capturing variants of some regexps and used that to reduce the
number of captured groups from 11 to 8.