sys-apps/systemd: Backport fix for networkd unmanaged issue to v256
This is upstream issue https://github.com/systemd/systemd/issues/36997 which we
are hitting on Auzre, most likely since the networking rules shipped by
azure-vm-utils are taking priority over Flatcar's own rule to keep enslaved
NICs unmanaged. The azure-vm-utils configs use the ID_NET_MANAGED_BY property
in addition to a custom property set on interfaces to keep them unmanaged. It
isn't yet clear if only ID_NET_MANAGED_BY is affected or if it's the only thing
that was fixed.

Meanwhile we need to backport the bugfix. Patch 0011 is the actual fix and 0010
is included because it makes the backport easier. The bugfix will be in v258
and has been backported to v257.7 upstream.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>