pkg-auto: Add a "bunch of maps" struct
The purpose of this struct is to collect all the information that is
needed for handling package updates in one place. It is not really
used right now, but when the package handling is split off into a
separate function, it will come in handy as we can then pass a couple
of parameters to the new function instead of many.

Also, in future the struct will grow, when we add ignoring irrelevant
information in summary stubs or license filtering.