sys-kernel/coreos-firmware: Partially resync with Gentoo
KV_FULL wasn't being set before, so it was actually scanning all modules
under /lib/modules, not just the ones for the current kernel version.

Hopefully, we can eventually create an env file so that we can use the
upstream ebuild directly, but this at least brings them closer in line.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>