sys-kernel/coreos-kernel: Decompress the minimal initrd kernel modules
The initrd itself is already compressed with xz, so compressing the
modules within with xz as well just makes the image bigger and slower.

Not compressing all the kernel modules in the first place is not such a
clear win because it is zlib and zstd compression that is applied
elsewhere.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>