ci-automation/release: Set up rclone config for mapping into container
The wrong variable was used to set up the mapping of the rclone config
into the rclone container and it wasn't set up in the right function.
Move it into the right function and use the right variable name but also
don't rely on /proc/PID/fd/FD to be mappable into the container but
instead use a regular temp file.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>