dev-db/etcd: Replace our etcdctl package with Gentoo's etcd package
A `server` USE flag avoids including that binary. The client tools now
include etcdutl for performing defragmentation and snapshot restoration
on local data.

We previously built etcdctl with `CGO_ENABLED=0`, apparently for
consistency when cross-compiling, but we have since fixed that.

Gentoo's package is not currently stabilised for arm64, as I have only
just added that keyword while bumping to the latest release. The latest
stable release is a little old, but still much newer than ours.

Closes: https://github.com/flatcar/Flatcar/issues/1932
Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>