dev-db/etcd: Replace our etcdctl package with Gentoo's etcd package
A `server` USE flag avoids including that binary. The client tools now
include etcdutl for performing defragmentation and snapshot restoration
on local data.

We previously built etcdctl with `CGO_ENABLED=0`, apparently for
consistency when cross-compiling, but we have since fixed that.

Gentoo's package is not currently keyworded for arm64, but we can
address that later. It's also a little outdated, but still much newer
than ours, and we can address that later too.

Closes: https://github.com/flatcar/Flatcar/issues/1932
Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>