app-admin/google-guest-configs: New package to install udev disk rules
We already have GCE disk rules in coreos-init, but a user has pointed
out that the newer NVMe rules are missing. Let's take the rules directly
from upstream instead. This is loosely based on the ChromiumOS package
of the same name.

This only installs the udev disk rules because the network rules are not
GCE-specific. The disk rules need to be in the initrd, so we cannot
separate this package from other platforms by putting it in the GCE OEM
image. We could split this package in two, but the network rules don't
seem essential.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>